# swagger.yaml
openapi: 3.0.3
info:
  title: Backup Tools API
  description: |
    A comprehensive backup and synchronization API that allows users to backup data from various cloud services 
    (Google Drive, Google Photos, Gmail, Google Cloud Storage, Dropbox, Office 365, AWS S3, GitHub, Shopify, QuickBooks) 
    to a satellite storage system and vice versa.
    
    ## Features
    - **Multi-Platform Support**: Integrate with major cloud providers
    - **Automatic Backup**: Schedule and manage automated backup jobs
    - **Satellite Storage**: Store and retrieve data from satellite storage
    - **Real-time Sync**: Live synchronization capabilities
    - **Authentication**: JWT-based authentication with OAuth2 flows
    
    ## Authentication
    Most endpoints require authentication via JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    Some endpoints also require an ACCESS_TOKEN header for satellite operations:
    ```
    ACCESS_TOKEN: <your-satellite-access-token>
    ```
  version: 1.0.0
  contact:
    name: Backup Tools Support
    email: support@backuptools.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8005
    description: Development server
  - url: https://backup.storx.io
    description: Production server
...
paths:
...
  /users-groups/mailbox/services:
    get:
      tags:
        - Auto Sync
      summary: Mailbox Services tab
      description: |
        All five Google workspace services with connected flag. Connect via POST /auto-sync/job onboarding.
      parameters:
        - $ref: '#/components/parameters/tokenKeyHeader'
        - name: email
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Mailbox services
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersGroupsMailboxServicesResponse'
...
  /auto-sync/job/services:
    get:
      tags:
        - Auto Sync
      summary: List All Autosync Services
      description: |
        Returns all supported autosync services with per-method job counts (Services Update page).
        Services with no jobs for the user return zero counts. Use filter.method on GET /auto-sync/job for Configure modal job listing.
      parameters:
        - $ref: '#/components/parameters/tokenKeyHeader'
      responses:
        '200':
          description: Connected services with per-method job counts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutosyncServicesResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    tokenKeyHeader:
      name: token_key
      in: header
      required: true
      schema:
        type: string
      description: Satellite session token (_tokenKey). Required for /auto-sync/*, /autosync/*, and /restore/* (Satellite BFF).
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"
        message:
          type: string
          example: "Error description"
...
    UsersGroupsMailboxHeader:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        account_type:
          type: string
          enum: [corporate, individual]
...
    UsersGroupsMailboxServicesItem:
      type: object
      properties:
        method:
          type: string
        connected:
          type: boolean
        job_id:
          type: integer
    UsersGroupsMailboxServicesResponse:
      allOf:
        - $ref: '#/components/schemas/UsersGroupsMailboxHeader'
        - type: object
          properties:
            services:
              type: array
              items:
                $ref: '#/components/schemas/UsersGroupsMailboxServicesItem'
...
    ServiceJobCountView:
      type: object
      properties:
        method:
          type: string
          enum: [gmail, google_drive, google_photos, google_contacts, google_calendar]
        total_jobs:
          type: integer
        active_jobs:
          type: integer
        deactive_jobs:
          type: integer
    AutosyncServicesResponse:
      type: object
      properties:
        message:
          type: string
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceJobCountView'
...
security:
  - bearerAuth: []